feat: As a Developer I want to add the capability of sharing operational monitoring data with the rest of the components via gRPC to the operational monitoring addon so that we can use it to visualise traffic
move common_messages.proto to common-rpc

refs: XRDDEV-2892, XRDDEV-2893
