+++ src/common/common-rpc/build.gradle.kts
@@ -17,6 +17,7 @@ dependencies {
   implementation(project(":common:common-core"))
   implementation(libs.slf4j.api)
 
+  api(project(":common:common-domain"))
   api(libs.grpc.protobuf)
   api(libs.grpc.stub)
   api(libs.grpc.nettyShaded)
