+++ src/service/signer/signer-core/src/main/java/org/niis/xroad/signer/core/protocol/handler/GenerateCertReqReqHandler.java
@@ -30,7 +30,7 @@ import ee.ria.xroad.common.CodedException;
 import com.google.protobuf.ByteString;
 import lombok.extern.slf4j.Slf4j;
 import org.bouncycastle.pkcs.PKCS10CertificationRequest;
-import org.niis.xroad.signer.api.mapper.ClientIdMapper;
+import org.niis.xroad.common.rpc.mapper.ClientIdMapper;
 import org.niis.xroad.signer.core.tokenmanager.TokenManager;
 import org.niis.xroad.signer.core.tokenmanager.token.SoftwareTokenType;
 import org.niis.xroad.signer.core.util.TokenAndKey;
