+++ src/service/signer/signer-core/src/main/java/org/niis/xroad/signer/core/protocol/handler/GetMemberCertsReqHandler.java
@@ -27,8 +27,8 @@ package org.niis.xroad.signer.core.protocol.handler;
 
 import ee.ria.xroad.common.identifier.ClientId;
 
+import org.niis.xroad.common.rpc.mapper.ClientIdMapper;
 import org.niis.xroad.signer.api.dto.CertificateInfo;
-import org.niis.xroad.signer.api.mapper.ClientIdMapper;
 import org.niis.xroad.signer.core.protocol.AbstractRpcHandler;
 import org.niis.xroad.signer.core.tokenmanager.TokenManager;
 import org.niis.xroad.signer.proto.GetMemberCertsReq;
